move inherited-value strategy as separate feature
