+++ src/app/core/common-components/entity-field-edit/entity-field-edit.component.ts
@@ -13,7 +13,7 @@ import {
 } from "../entity-form/FormConfig";
 import { NgClass, NgIf } from "@angular/common";
 import { EntityFieldViewComponent } from "../entity-field-view/entity-field-view.component";
-import { InheritedValueButtonComponent } from "../../default-values/x-inherited-value/inherited-value-button/inherited-value-button.component";
+import { InheritedValueButtonComponent } from "../../../features/default-value-inherited/inherited-value-button/inherited-value-button.component";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { MatButtonModule } from "@angular/material/button";
 import { MatTooltipModule } from "@angular/material/tooltip";
