+++ src/app/core/default-values/standard-default-value-strategies.ts
@@ -1,7 +1,7 @@
 import { DefaultValueStrategy } from "./default-value-strategy.interface";
 import { StaticDefaultValueService } from "./x-static/static-default-value.service";
 import { DynamicPlaceholderValueService } from "./x-dynamic-placeholder/dynamic-placeholder-value.service";
-import { InheritedValueService } from "./x-inherited-value/inherited-value.service";
+import { InheritedValueService } from "../../features/default-value-inherited/inherited-value.service";
 
 /**
  * Standard default-value strategies that are used in the application.
