+++ src/app/features/default-value-inherited/admin-default-value-inherited/admin-default-value-inherited.component.scss
