+++ src/app/features/default-value-inherited/default-value-config-inherited.ts
