+++ src/app/features/default-value-inherited/inherited-value-button/inherited-value-button.component.scss
