+++ src/app/features/default-value-inherited/inherited-value.service.spec.ts
@@ -1,21 +1,21 @@
 import { fakeAsync, TestBed, tick } from "@angular/core/testing";
 
 import { InheritedValueService } from "./inherited-value.service";
-import { Entity } from "../../entity/model/entity";
-import { EntityMapperService } from "../../entity/entity-mapper/entity-mapper.service";
+import { Entity } from "../../core/entity/model/entity";
+import { EntityMapperService } from "../../core/entity/entity-mapper/entity-mapper.service";
 import {
   cleanUpTemporarySchemaFields,
   getDefaultInheritedForm,
-} from "../default-value-service/default-value.service.spec";
+} from "../../core/default-values/default-value-service/default-value.service.spec";
 import { FormControl, FormGroup } from "@angular/forms";
-import { EntityForm } from "../../common-components/entity-form/entity-form.service";
-import { DefaultValueService } from "../default-value-service/default-value.service";
+import { EntityForm } from "../../core/common-components/entity-form/entity-form.service";
+import { DefaultValueService } from "../../core/default-values/default-value-service/default-value.service";
 import { EventEmitter } from "@angular/core";
-import { EntityAbility } from "../../permissions/ability/entity-ability";
-import { UpdatedEntity } from "../../entity/model/entity-update";
-import { Config } from "../../config/config";
+import { EntityAbility } from "../../core/permissions/ability/entity-ability";
+import { UpdatedEntity } from "../../core/entity/model/entity-update";
+import { Config } from "../../core/config/config";
 import { Subject } from "rxjs";
-import { DefaultValueStrategy } from "../default-value-strategy.interface";
+import { DefaultValueStrategy } from "../../core/default-values/default-value-strategy.interface";
 
 describe("InheritedValueService", () => {
   let service: InheritedValueService;
